refactor: normalize intra-file ordering and group module init
Apply a consistent top-down structure across the touched files:
helpers/types hoisted above their consumers, exported interfaces placed
before their implementations, module-init side effects (BroadcastChannel
listener, voiceschanged listener) sit next to the stores they wire up,
and lifecycle cleanups land at the bottom of hooks. db.ts gains section
comment dividers grouping records / inputs / schema / connection /
boardSets / boards / assets. Extract `makeChunkStream` from
ai-namespace-fake.ts into a dedicated stream-fake.ts file so neither
side carries a misleading import.

Behavior unchanged.
