+++ src/features/board/board-set/info-dialog.tsx
@@ -15,33 +15,33 @@ export interface BoardSetInfoDialogProps {
   onClose: () => void;
 }
 
+function buildChipLabels(boardSet: BoardSetRecord): string[] {
+  const labels: string[] = [];
+
+  if (boardSet.gridRows && boardSet.gridColumns) {
+    labels.push(
+      m.libraryGridDimensions({
+        rows: boardSet.gridRows,
+        columns: boardSet.gridColumns,
+      }),
+    );
+  }
+
+  if (boardSet.locale) {
+    labels.push(getEnglishLocaleName(boardSet.locale));
+  }
+
+  if (boardSet.license) {
+    labels.push(boardSet.license);
+  }
+
+  return labels;
+}
+
 export function BoardSetInfoDialog({
   boardSet,
   onClose,
 }: BoardSetInfoDialogProps) {
-  function buildChipLabels(boardSet: BoardSetRecord): string[] {
-    const labels: string[] = [];
-
-    if (boardSet.gridRows && boardSet.gridColumns) {
-      labels.push(
-        m.libraryGridDimensions({
-          rows: boardSet.gridRows,
-          columns: boardSet.gridColumns,
-        }),
-      );
-    }
-
-    if (boardSet.locale) {
-      labels.push(getEnglishLocaleName(boardSet.locale));
-    }
-
-    if (boardSet.license) {
-      labels.push(boardSet.license);
-    }
-
-    return labels;
-  }
-
   const chipLabels = boardSet ? buildChipLabels(boardSet) : [];
 
   return (
