+++ src/features/board/storage/queries.ts
@@ -22,6 +22,11 @@ export class BoardNotFoundError extends Error {
   }
 }
 
+// Single concurrent caller assumed — the only consumer is boardLoader.
+// Revoke on the next call rather than from the caller: the consumer can't
+// know when its URLs are safe to release; the next load defines that boundary.
+let previousRegistry: ObjectUrlRegistry | null = null;
+
 export async function getBoardSet(
   setId: string,
 ): Promise<BoardSetRecord | undefined> {
@@ -38,11 +43,6 @@ export async function persistBoardTranslations(
     updateBoardStrings(db, setId, boardId, locale, translations),
   );
 }
-
-// Single concurrent caller assumed — the only consumer is boardLoader.
-// Revoke on the next call rather than from the caller: the consumer can't
-// know when its URLs are safe to release; the next load defines that boundary.
-let previousRegistry: ObjectUrlRegistry | null = null;
 
 export async function hydrateBoard(
   setId: string,
