+++ src/features/board/types.ts
@@ -43,12 +43,6 @@ export type BoardAction =
   | { kind: "speak" }
   | { kind: "spell"; text: string };
 
-export function getSpokenText(
-  button: Pick<BoardButton, "vocalization" | "label">,
-): string | undefined {
-  return button.vocalization ?? button.label;
-}
-
 export interface BoardLicense {
   type: string;
   authorName?: string;
@@ -59,3 +53,9 @@ export interface BoardLicense {
 }
 
 export type BoardStrings = Record<string, Record<string, string>>;
+
+export function getSpokenText(
+  button: Pick<BoardButton, "vocalization" | "label">,
+): string | undefined {
+  return button.vocalization ?? button.label;
+}
