+++ src/shared/built-in-ai/internal/testing/ai-namespace-fake.ts
@@ -39,17 +39,3 @@ export function makeAIFake<I>({
     instances,
   };
 }
-
-/** `ReadableStream<string>` that emits `chunks` in order then closes. */
-export function makeChunkStream(
-  chunks: readonly string[],
-): ReadableStream<string> {
-  return new ReadableStream<string>({
-    start(controller) {
-      for (const c of chunks) {
-        controller.enqueue(c);
-      }
-      controller.close();
-    },
-  });
-}
