+++ src/shared/built-in-ai/internal/testing/instance-fakes.ts
@@ -1,27 +1,5 @@
 import { vi } from "vitest";
-import { makeChunkStream } from "./ai-namespace-fake";
-
-export function buildTranslatorInstance() {
-  return {
-    translate: vi.fn((input: string) => Promise.resolve(`T:${input}`)),
-    translateStreaming: vi.fn(() => makeChunkStream(["T:", "hello"])),
-    measureInputUsage: vi.fn(() => Promise.resolve(7)),
-    inputQuota: 1024,
-    destroy: vi.fn<() => void>(),
-  };
-}
-
-export function buildRewriterInstance() {
-  return {
-    rewrite: vi.fn((input: string, opts?: RewriterRewriteOptions) =>
-      Promise.resolve(`R(${opts?.context ?? ""}):${input}`),
-    ),
-    rewriteStreaming: vi.fn(() => makeChunkStream(["R:", "alt"])),
-    measureInputUsage: vi.fn(() => Promise.resolve(4)),
-    inputQuota: 768,
-    destroy: vi.fn<() => void>(),
-  };
-}
+import { makeChunkStream } from "./stream-fake";
 
 export function buildProofreaderInstance() {
   return {
@@ -38,6 +16,28 @@ export function buildProofreaderInstance() {
         ],
       }),
     ),
+    destroy: vi.fn<() => void>(),
+  };
+}
+
+export function buildRewriterInstance() {
+  return {
+    rewrite: vi.fn((input: string, opts?: RewriterRewriteOptions) =>
+      Promise.resolve(`R(${opts?.context ?? ""}):${input}`),
+    ),
+    rewriteStreaming: vi.fn(() => makeChunkStream(["R:", "alt"])),
+    measureInputUsage: vi.fn(() => Promise.resolve(4)),
+    inputQuota: 768,
+    destroy: vi.fn<() => void>(),
+  };
+}
+
+export function buildTranslatorInstance() {
+  return {
+    translate: vi.fn((input: string) => Promise.resolve(`T:${input}`)),
+    translateStreaming: vi.fn(() => makeChunkStream(["T:", "hello"])),
+    measureInputUsage: vi.fn(() => Promise.resolve(7)),
+    inputQuota: 1024,
     destroy: vi.fn<() => void>(),
   };
 }
